+++ src/screens/Home/Views/Setting/settings/Backup/Part.tsx
@@ -7,11 +7,13 @@ import SubTitle from '../../components/SubTitle'
 import Button from '../../components/Button'
 import { useI18n } from '@/lang'
 import ListImportExport, { type ListImportExportType } from './ListImportExport'
+import SettingsImportExport, { type SettingsImportExportType } from './SettingsImportExport'
 
 
 export default memo(() => {
   const t = useI18n()
   const listImportExportRef = useRef<ListImportExportType>(null)
+  const settingsImportExportRef = useRef<SettingsImportExportType>(null)
 
   return (
     <>
@@ -19,17 +21,16 @@ export default memo(() => {
         <View style={styles.list}>
           <Button onPress={() => listImportExportRef.current?.import()}>{t('setting_backup_part_import_list')}</Button>
           <Button onPress={() => listImportExportRef.current?.export()}>{t('setting_backup_part_export_list')}</Button>
-          {/* <Button onPress={() => importAndExportData('import', 'setting')}>{t('setting_backup_part_import_setting')}</Button>
-          <Button onPress={() => importAndExportData('export', 'setting')}>{t('setting_backup_part_export_setting')}</Button> */}
         </View>
       </SubTitle>
-      {/* <SubTitle title={t('setting_backup_all')}>
+      <SubTitle title={t('setting_backup_all')}>
         <View style={styles.list}>
-          <Button onPress={() => importAndExportData('import', 'all')}>{t('setting_backup_all_import')}</Button>
-          <Button onPress={() => importAndExportData('export', 'all')}>{t('setting_backup_all_export')}</Button>
+          <Button onPress={() => settingsImportExportRef.current?.import()}>{t('setting_backup_part_import_setting')}</Button>
+          <Button onPress={() => settingsImportExportRef.current?.export()}>{t('setting_backup_part_export_setting')}</Button>
         </View>
-      </SubTitle> */}
+      </SubTitle>
       <ListImportExport ref={listImportExportRef} />
+      <SettingsImportExport ref={settingsImportExportRef} />
     </>
   )
 })